20 delete slot dialog

+++ Bezahlsystem/OAOPS/Client/Pages/AdminArea/SlotManagement.razor.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Rendering;
+using MudBlazor;
 using OAOPS.Client.Components.AddComponents;
+using OAOPS.Client.Components.Shared;
 using OAOPS.Client.Components.UpdateComponents;
 using OAOPS.Client.DTO;
 using OAOPS.Client.Services;
@@ -31,16 +33,19 @@ namespace OAOPS.Client.Pages.AdminArea
         [Inject]
         public IDialogService DialogService { get; set; }
 
-        public List<StorageSlotDto>? Slots { get; set; }
+        [Inject]
+        public ISnackbar Snackbar { get; set; }
+
+        public List<StorageSlotDto>? Slots { get; set; } = new();
 
         protected override async Task OnInitializedAsync()
         {
-            Slots = await dataService.GetSlotsOfStorageByName(Name) ?? new();
+            await LoadSlots();
+        }
 
-            foreach (StorageSlotDto slot in Slots)
-            {
-                await Console.Out.WriteLineAsync($"{slot.SlotId} {slot.SlotName} {slot.StorageConnectionId} {slot.ArticleName}");
-            }
+        private async Task LoadSlots()
+        {
+            Slots = await dataService.GetSlotsOfStorageByName(Name) ?? new();
         }
 
         void GoBack()
@@ -50,20 +55,41 @@ namespace OAOPS.Client.Pages.AdminArea
 
         async Task DeleteSlot(StorageSlotDto slot)
         {
-            await dataService.DeleteStorageSlot(slot.SlotId);
+            var param = new DialogParameters
+            {
+                { "Name", slot.SlotName + " Slot" }
+            };
+            var res = await DialogService.Show<DeleteConfirmationDialog>("Delete Slot", param, new DialogOptions() { CloseButton = true, MaxWidth = MaxWidth.Medium }).Result;
+            if (res.Canceled)
+            {
+                return;
+            }
+
+            var error = await dataService.DeleteStorageSlot(slot.SlotId);
+            if (error.IsSuccessCode)
+            {
+                Snackbar.Add($"Slot {slot.SlotName} erfolgreich gelöscht.", Severity.Success);
+            }
+            else
+            {
+                Snackbar.Add($"Fehler beim löschen des Slots {slot.SlotName}", Severity.Error);
+            }
+
+            await LoadSlots();
         }
 
-        void AddSlot()
+        async Task AddSlot()
         {
             var parameter = new DialogParameters()
             {
                 {"StorageName", Name }
             };
 
-            DialogService.Show<AddStorageSlot>("Add Slot", parameter ,new DialogOptions() { CloseButton = true, MaxWidth = MaxWidth.Medium });
+            var result = await DialogService.Show<AddStorageSlot>("Add Slot", parameter, new DialogOptions() { CloseButton = true, MaxWidth = MaxWidth.Medium }).Result;
+            await LoadSlots();
         }
 
-        void UpdateSlot(StorageSlotDto slot)
+        async Task UpdateSlot(StorageSlotDto slot)
         {
             var param = new DialogParameters
             {
@@ -71,7 +97,9 @@ namespace OAOPS.Client.Pages.AdminArea
             };
 
             var opt = new DialogOptions() { CloseButton = true, MaxWidth = MaxWidth.Medium };
-            DialogService.Show<UpdateStorageSlot>("Slot Update", param, opt);
+            await DialogService.Show<UpdateStorageSlot>("Slot Update", param, opt).Result;
+
+            await LoadSlots();
         }
     }
 }
